Fix/update `getUsersConfirmedButNotSettledSynthBalance` so it doesn't include tokens already transferred away from the user.

+++ contracts/contracts/SyntheticToken.sol
@@ -78,13 +78,11 @@ contract SyntheticToken is ISyntheticToken {
    */
   function balanceOf(address account) public view virtual override returns (uint256) {
     return
-      uint256(
-        int256(ERC20.balanceOf(account)) +
-          ILongShort(longShort).getUsersConfirmedButNotSettledSynthBalance(
-            account,
-            marketIndex,
-            isLong
-          )
+      ERC20.balanceOf(account) +
+      ILongShort(longShort).getUsersConfirmedButNotSettledSynthBalance(
+        account,
+        marketIndex,
+        isLong
       );
   }
 }
